chore: move pages to the retail app package
