+++ packages/adyen-retail-react-app/overrides/app/pages/checkout/partials/payment.jsx
@@ -22,8 +22,8 @@ import AddressDisplay from '@salesforce/retail-react-app/app/components/address-
 import {PromoCode, usePromoCode} from '@salesforce/retail-react-app/app/components/promo-code'
 import {API_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/constants'
 import {useCheckout} from '@salesforce/retail-react-app/app/pages/checkout/util/checkout-context'
-import AdyenCheckout from '../../../components/adyenCheckout'
-import {useAdyenCheckout} from '../../../context/adyen-checkout-context'
+import AdyenCheckout from '@adyen/adyen-salesforce-pwa/lib/components/adyenCheckout'
+import {useAdyenCheckout} from '@adyen/adyen-salesforce-pwa/lib/context/adyen-checkout-context'
 
 const Payment = ({useShopperBasketsMutation}) => {
     const {formatMessage} = useIntl()
